Refactor #18 - Routes에서 Controller로 기능 이전
